pyproject.toml
[tool.poetry]
name = "br2-vision-based-smoothing"
version = "0.2.3"
description = "Video tracking and strain reconstruction algorithm for soft-arm project."
authors = ["Seung Hyun Kim <skim449@illinois.edu>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "br2_vision" },
    { include = "script" },
]

[tool.poetry.scripts]
undistort_rotate_video = 'script.undistort_rotate_video:undistort_and_rotate'
sync_video = 'script.sync_video:process'
crop_video = 'script.crop_video:process'
trim_video_intervals_by_led = 'script.trim_video_intervals_by_led:process'
trim_video_intervals_manual = 'script.trim_video_intervals_manual:process'

select_calibration_points = 'script.run_select_calibration_points:select_calibration_points'
extract_frames_from_video = 'script.extract_frames_from_video:extract_frames'
dlt_calibration = 'script.dlt_calibration:calibrate'

distortion_calibration = 'script.distortion_calibration:calibrate'

set_optical_flow_inquiry = 'script.add_initial_flow_point:main'
run_optical_flow = 'script.run_optical_flow:main'
edit_optical_flow = 'script.interactive_cutout_trajectory:launch'

interpolate_poses = 'script.process_dlt:process_dlt'

overlay_framemark = 'script.overlay_framemark:process'

[tool.poetry.dependencies]
python = "^3.9,<3.12"
pyelastica = "^0.3.1.post1"
opencv-python = "^4.0"
tqdm = "^4.66.1"
scikit-learn = "^1.3.0"
scikit-spatial = "^7.0.0"
scipy = "^1.11.2"
matplotlib = "^3.7.2"
click = "^8.1.7"
psutil = "^5.9.6"
pyyaml = "^6.0.1"
pyqt6 = "^6.6.1"
dash = {extras = ["diskcache"], version = "^2.15.0"}
diskcache = "^5.6.3"
dash-bootstrap-components = "^1.5.0"
pandas = "^2.2.0"
h5py = "^3.10.0"
coverage = { version = "^7.4.4", extras = ["toml"] }

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
# Regexes for lines to exclude from consideration
exclude_lines = [
    # Enable pragma
    "pragma: no cover",
    "TODO",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == __main__:",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    '''class '.*\bProtocol\)':''',
    ''''@(abc\.)?'abstractmethod''',
    '\.\.\.'
]
#fail_under = 40
show_missing = true


[tool.coverage.run]
branch = true
omit = [
    "*/.local/*",
	"setup.py",
    "*/__init__.py",
    "*/tests/*",
]
